code/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
crewai==0.1.7
langchain==0.0.351
openai==0.28.1
langchain-community==0.0.13
upwork
psycopg2-binary
chromadb==0.4.22
tokenizers==0.13.2
langchain-core
langchain-openai